Type cast width to short to avoid type cast exception in debug version when width is negative.

+++ xorg-server/miext/damage/damage.c
@@ -1331,9 +1331,9 @@ damagePolyFillRect(DrawablePtr	pDrawable,
 	int	    nRectsTmp = nRects;
 
 	box.x1 = pRectsTmp->x;
-	box.x2 = box.x1 + pRectsTmp->width;
+	box.x2 = box.x1 + (short)pRectsTmp->width;
 	box.y1 = pRectsTmp->y;
-	box.y2 = box.y1 + pRectsTmp->height;
+	box.y2 = box.y1 + (short)pRectsTmp->height;
 
 	while(--nRectsTmp) 
 	{
